+++ Content.Shared/_CE/EntityEffect/Effects/ApplyStatusEffectStack.cs
@@ -1,3 +1,4 @@
+using Content.Shared._CE.StatusEffects.Core;
 using Content.Shared._CE.StatusEffectStacks;
 using Robust.Shared.Prototypes;
