+++ Content.Shared/_CE/Equipment/CEEquipStatusEffectSystem.cs
@@ -1,3 +1,4 @@
+using Content.Shared._CE.StatusEffects.Core;
 using Content.Shared._CE.StatusEffectStacks;
 using Content.Shared.Inventory.Events;
