+++ Content.Shared/_CE/Skill/Skills/EffectsOnTriggerStatusEffects/CEEffectsOnTriggerStatusEffectSystem.cs
@@ -2,6 +2,7 @@ using Content.Shared._CE.EntityEffect;
 using Content.Shared._CE.Health;
 using Content.Shared._CE.MeleeWeapon;
 using Content.Shared._CE.Skill.Skills.EffectsOnTriggerStatusEffects.Components;
+using Content.Shared._CE.StatusEffects.Core;
 using Content.Shared._CE.StatusEffectStacks;
 using Content.Shared.StatusEffectNew;
 using Content.Shared.StatusEffectNew.Components;
