+++ Content.Shared/_CE/Skill/Skills/Wisdom/CEWisdomStatusEffectSystem.cs
@@ -1,4 +1,5 @@
 using Content.Shared._CE.Mana.Core;
+using Content.Shared._CE.StatusEffects.Core;
 using Content.Shared._CE.StatusEffectStacks;
 using Content.Shared.StatusEffectNew;
 using Content.Shared.StatusEffectNew.Components;
