+++ Content.Shared/_CE/StatusEffects/Regeneration/CERegenerationStatusEffectSystem.cs
@@ -1,4 +1,5 @@
 using Content.Shared._CE.Health;
+using Content.Shared._CE.StatusEffects.Core;
 using Content.Shared._CE.StatusEffectStacks;
 using Content.Shared.StatusEffectNew.Components;
