+++ Content.Shared/_CE/StatusEffects/StaminaRegenBuff/CEStaminaRegenBuffSystem.cs
@@ -1,4 +1,5 @@
 using Content.Shared._CE.Stamina;
+using Content.Shared._CE.StatusEffects.Core;
 using Content.Shared._CE.StatusEffectStacks;
 using Content.Shared.StatusEffectNew;
